# implemented approximate gadget decomposition + optimization for LMKCDEY