--- a/src/binfhe/lib/rgsw-acc.cpp
+++ b/src/binfhe/lib/rgsw-acc.cpp
@@ -58,7 +58,8 @@ void RingGSWAccumulator::SignedDigitDecompose(const std::shared_ptr<RingGSWCrypt
     auto Q_int{params->GetQ().ConvertToInt<NativeInteger::SignedNativeInt>()};
     auto gBits{static_cast<NativeInteger::SignedNativeInt>(__builtin_ctz(params->GetBaseG()))};
     auto gBitsMaxBits{static_cast<NativeInteger::SignedNativeInt>(NativeInteger::MaxBits() - gBits)};
-    uint32_t digitsG2{params->GetDigitsG() << 1};
+    // approximate gadget decomposition is used; the first digit is ignored
+    uint32_t digitsG2{(params->GetDigitsG() - 1) << 1};
     uint32_t N{params->GetN()};
 
     for (uint32_t k{0}; k < N; ++k) {
@@ -67,14 +68,20 @@ void RingGSWAccumulator::SignedDigitDecompose(const std::shared_ptr<RingGSWCrypt
         auto t1{input[1][k].ConvertToInt<BasicInteger>()};
         auto d1{static_cast<NativeInteger::SignedNativeInt>(t1 < QHalf ? t1 : t1 - Q_int)};
 
+        auto r0{(d0 << gBitsMaxBits) >> gBitsMaxBits};
+        d0 = (d0 - r0) >> gBits;
+
+        auto r1{(d1 << gBitsMaxBits) >> gBitsMaxBits};
+        d1 = (d1 - r1) >> gBits;
+
         for (uint32_t d{0}; d < digitsG2; d += 2) {
-            auto r0{(d0 << gBitsMaxBits) >> gBitsMaxBits};
+            r0 = (d0 << gBitsMaxBits) >> gBitsMaxBits;
             d0 = (d0 - r0) >> gBits;
             if (r0 < 0)
                 r0 += Q_int;
             output[d + 0][k] += r0;
 
-            auto r1{(d1 << gBitsMaxBits) >> gBitsMaxBits};
+            r1 = (d1 << gBitsMaxBits) >> gBitsMaxBits;
             d1 = (d1 - r1) >> gBits;
             if (r1 < 0)
                 r1 += Q_int;
@@ -85,21 +92,24 @@ void RingGSWAccumulator::SignedDigitDecompose(const std::shared_ptr<RingGSWCrypt
 
 // Decompose a ring element, not ciphertext
 void RingGSWAccumulator::SignedDigitDecompose(const std::shared_ptr<RingGSWCryptoParams>& params,
-                                              const NativePoly& input,
-                                              std::vector<NativePoly>& output) const {
+                                              const NativePoly& input, std::vector<NativePoly>& output) const {
     auto QHalf{params->GetQ().ConvertToInt<BasicInteger>() >> 1};
     auto Q_int{params->GetQ().ConvertToInt<NativeInteger::SignedNativeInt>()};
     auto gBits{static_cast<NativeInteger::SignedNativeInt>(__builtin_ctz(params->GetBaseG()))};
     auto gBitsMaxBits{static_cast<NativeInteger::SignedNativeInt>(NativeInteger::MaxBits() - gBits)};
-    uint32_t digitsG{params->GetDigitsG()};
+    // approximate gadget decomposition is used; the first digit is ignored
+    uint32_t digitsG{params->GetDigitsG() - 1};
     uint32_t N{params->GetN()};
 
     for (uint32_t k{0}; k < N; ++k) {
         auto t0{input[k].ConvertToInt<BasicInteger>()};
         auto d0{static_cast<NativeInteger::SignedNativeInt>(t0 < QHalf ? t0 : t0 - Q_int)};
 
+        auto r0{(d0 << gBitsMaxBits) >> gBitsMaxBits};
+        d0 = (d0 - r0) >> gBits;
+
         for (uint32_t d{0}; d < digitsG; ++d) {
-            auto r0{(d0 << gBitsMaxBits) >> gBitsMaxBits};
+            r0 = (d0 << gBitsMaxBits) >> gBitsMaxBits;
             d0 = (d0 - r0) >> gBits;
             if (r0 < 0)
                 r0 += Q_int;