* Re-implemented the changes in PR #482 as the binfhe code has significantly changed as compared to the commit the branch was created from (it was easier to reimplement than resolve the conflicts)
* Optimized the `Automorphism` bootstrapping step for the LMKCDEY mode
    * Removed 2 NTTs (automorphisms can be done directly in EVALUATION representation)
    * Switched to an optimized version of automorphisms (with precomputed bit reversal tables)

Ran benchmarks for STD128* on an Intel(R) Core(TM) i7-9700 CPU @ 3.00GHz and 64 GB of RAM, running Ubuntu 20.04 LTS with clang 10.0.0 (CMake flags: NATIVE_SIZE=32; WITH_NATIVEOPT=ON; OMP_NUM_THREADS=1)
* GINX (TFHE) runtime went down from 49 ms to 39 ms.
* LMKCDEY (FHEW) runtime went down from 40 ms to 29 ms.