@yspolyakov 2023-07-14
The benchmark results for  CLANG 10.0.0; NATIVE_SIZE=32; WITH_NATIVEOPT=ON;  OMP_NUM_THREADS=1 
<table>
<tr>
<th>dev branch</th>
<th>modified_math branch</th>
</tr>
<tr>
<td>

```
 bin/benchmark/binfhe-ginx
2023-07-13T17:22:08-04:00
Running bin/benchmark/binfhe-ginx
Run on (8 X 4707.17 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 256 KiB (x8)
  L3 Unified 12288 KiB (x1)
Load Average: 4.32, 2.48, 1.40
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
FHEW_BTKEYGEN/MEDIUM                     31079099 us       885542 us            1
FHEW_BTKEYGEN/STD128                      3041986 us      1220569 us            1
FHEW_ENCRYPT/MEDIUM                          16.7 us         16.7 us        41987
FHEW_ENCRYPT/STD128                          20.2 us         20.2 us        34697
FHEW_NOT/MEDIUM                             0.098 us        0.098 us      7115550
FHEW_NOT/STD128                             0.099 us        0.099 us      6941123
FHEW_BINGATE/MEDIUM_OR/min_time:10.000      31910 us        31909 us          438
FHEW_BINGATE/MEDIUM_AND                     32054 us        32053 us           22
FHEW_BINGATE/MEDIUM_NOR                     32168 us        32167 us           22
FHEW_BINGATE/MEDIUM_NAND                    32122 us        32122 us           22
FHEW_BINGATE/MEDIUM_XOR                     96653 us        96651 us            7
FHEW_BINGATE/MEDIUM_XNOR                    97045 us        97039 us            7
FHEW_BINGATE/MEDIUM_XOR_FAST                31979 us        31978 us           22
FHEW_BINGATE/MEDIUM_XNOR_FAST               32072 us        32071 us           22
FHEW_BINGATE/STD128_OR/min_time:10.000      48906 us        48905 us          284
FHEW_BINGATE/STD128_AND                     49009 us        49008 us           14
FHEW_BINGATE/STD128_NOR                     49179 us        49179 us           14
FHEW_BINGATE/STD128_NAND                    49156 us        49154 us           14
FHEW_BINGATE/STD128_XOR                    148204 us       148196 us            5
FHEW_BINGATE/STD128_XNOR                   146888 us       146887 us            5
FHEW_BINGATE/STD128_XOR_FAST                49200 us        49199 us           14
FHEW_BINGATE/STD128_XNOR_FAST               49131 us        49130 us           14
FHEW_KEYSWITCH/MEDIUM/min_time:1.000          202 us          202 us         6844
FHEW_KEYSWITCH/STD128/min_time:1.000          225 us          225 us         5928
```

</td>
<td>

```
 bin/benchmark/binfhe-ginx
2023-07-13T17:22:08-04:00
Running bin/benchmark/binfhe-ginx
Run on (8 X 4707.17 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 256 KiB (x8)
  L3 Unified 12288 KiB (x1)
Load Average: 4.32, 2.48, 1.40
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
FHEW_BTKEYGEN/MEDIUM                     31079099 us       885542 us            1
FHEW_BTKEYGEN/STD128                      3041986 us      1220569 us            1
FHEW_ENCRYPT/MEDIUM                          16.7 us         16.7 us        41987
FHEW_ENCRYPT/STD128                          20.2 us         20.2 us        34697
FHEW_NOT/MEDIUM                             0.098 us        0.098 us      7115550
FHEW_NOT/STD128                             0.099 us        0.099 us      6941123
FHEW_BINGATE/MEDIUM_OR/min_time:10.000      31910 us        31909 us          438
FHEW_BINGATE/MEDIUM_AND                     32054 us        32053 us           22
FHEW_BINGATE/MEDIUM_NOR                     32168 us        32167 us           22
FHEW_BINGATE/MEDIUM_NAND                    32122 us        32122 us           22
FHEW_BINGATE/MEDIUM_XOR                     96653 us        96651 us            7
FHEW_BINGATE/MEDIUM_XNOR                    97045 us        97039 us            7
FHEW_BINGATE/MEDIUM_XOR_FAST                31979 us        31978 us           22
FHEW_BINGATE/MEDIUM_XNOR_FAST               32072 us        32071 us           22
FHEW_BINGATE/STD128_OR/min_time:10.000      48906 us        48905 us          284
FHEW_BINGATE/STD128_AND                     49009 us        49008 us           14
FHEW_BINGATE/STD128_NOR                     49179 us        49179 us           14
FHEW_BINGATE/STD128_NAND                    49156 us        49154 us           14
FHEW_BINGATE/STD128_XOR                    148204 us       148196 us            5
FHEW_BINGATE/STD128_XNOR                   146888 us       146887 us            5
FHEW_BINGATE/STD128_XOR_FAST                49200 us        49199 us           14
FHEW_BINGATE/STD128_XNOR_FAST               49131 us        49130 us           14
FHEW_KEYSWITCH/MEDIUM/min_time:1.000          202 us          202 us         6844
FHEW_KEYSWITCH/STD128/min_time:1.000          225 us          225 us         5928
```

</td>
</tr>
</table>